ajout commentaires

+++ src/classes/utilitaire/Avis.php
@@ -4,11 +4,11 @@ namespace netvod\utilitaire;
 
 class Avis{
 
-    private int $note;
+    private float $note;
     private string $commentaire;
     private string $nomUtilisateur;
 
-    public function __construct(int $note, string $commentaire, string $nomUtilisateur){
+    public function __construct(float $note, string $commentaire, string $nomUtilisateur){
         $this->note = $note;
         $this->commentaire = $commentaire;
         $this->nomUtilisateur = $nomUtilisateur;
